App/requirements.txt
azure-cognitiveservices-language-luis==0.7.0
azure-common==1.1.27
azure-core==1.17.0
azure-functions==1.7.2
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.1.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.0.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.8.1
azureml==0.2.7
azureml-core==1.33.0
azureml-dataprep==2.20.1
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==1.18.2
azureml-dataset-runtime==1.33.0
azureml-telemetry==1.33.0
azureml-widgets==1.33.0